Adding Knuth (1975) estimator to TrainerFactory and switching to it in bnb-depth-test.
git-svn-id: svn+ssh://external.hltcoe.jhu.edu/home/hltcoe/mgormley/public/repos/dep_parse_filtered/trunk@723 0408881f-cb4b-4750-8b71-c94bcb194269

+++ src/edu/jhu/hltcoe/gridsearch/randwalk/RandWalkBnbNodeSampler.java
@@ -70,7 +70,7 @@ public class RandWalkBnbNodeSampler extends DmvLazyBranchAndBoundSolver {
         public int maxSamples = 10000;
         // TODO: push these parameters into superclass.
         public double epsilon = 0.1;
-        public int timeoutSeconds = 100;
+        public double timeoutSeconds = 100;
         public DependencyParserEvaluator evaluator = null;
     }
